[Feat#177] UserInfoManager로 UserRole 관리 로직 이동 및 TokenManager 정리
