+++ app/src/main/java/com/kiero/core/localstorage/TokenManager.kt
@@ -8,8 +8,6 @@ interface TokenManager {
     suspend fun saveTokens(accessToken: String, refreshToken: String)
     suspend fun getAccessToken(): String?
     suspend fun getRefreshToken(): String?
+    suspend fun clearAllData()
     suspend fun clearTokens()
-
-    suspend fun saveUserRole(role: UserRole)
-    suspend fun getUserRole(): UserRole?
 }
